+++ api/outlookOAuthCallback.js
@@ -1,0 +1,4 @@
+"use strict";
+
+const { outlookOAuthCallback } = require("../server/outlook-runtime");
+module.exports = outlookOAuthCallback;