+++ api/outlookWebhook.js
@@ -1,0 +1,4 @@
+"use strict";
+
+const { outlookWebhook } = require("../server/outlook-runtime");
+module.exports = outlookWebhook;